% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_agg_panel}
\alias{dictionary_agg_panel}
\title{FlowRepository columns description}
\format{
data.frame
\describe{
\item{\code{Column}}{fr table columns names}
\item{\code{Description}}{fr table columns description}
\item{\code{Origin}}{fr table columns page type}}
}
\usage{
dictionary_agg_panel
}
\description{
Description of FlowRepository experiment table columns.
}
\keyword{datasets}
