% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fr_panels}
\alias{fr_panels}
\title{FlowRepository FCS associated with panel}
\format{
data.frame
\describe{
\item{\code{rep_id}}{flowrepository id}
\item{\code{exp_id}}{experiment id}
\item{\code{fcs_file}}{fcs file name}
\item{\code{panel}}{marker vector separated with pipes}}
}
\usage{
fr_panels
}
\description{
Table containing the name of each fcs file referenced for each FlowRepository ID and their associated panel.
}
\keyword{datasets}
