% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fr_tbl}
\alias{fr_tbl}
\title{Public experiment representations table}
\format{
data.frame
\describe{
\item{\code{Experiment.Name}}{Experiment comprehensive name}
\item{\code{Repository.ID}}{Flowrepository id}
\item{\code{Manuscript.s.}}{PMID vector}
\item{\code{Journal.s.}}{NAs, Published Journal name encoded with illustrations}
\item{\code{FCS.Files}}{Number of fcs}
\item{\code{Illustrations}}{Number of illustrations}
\item{\code{Primary.Researcher}}{Contact name}
\item{\code{Project}}{Project}
\item{\code{MIFlowCyt.Score}}{Nas, Percentage score}
\item{\code{Updated}}{Update date Month/Year}}
}
\usage{
fr_tbl
}
\description{
Public experience table export from FlowRepository on \code{2024/08/01}.
}
\keyword{datasets}
