% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fr_tbl_version}
\alias{fr_tbl_version}
\title{FlowRepository Table Version}
\format{
Character. A string in the format "YYMMDD"
}
\usage{
fr_tbl_version
}
\description{
A character string representing the version of the table extracted from the FlowRepository. The format follows a year/month/day convention.
}
\keyword{datasets}
